refactor: split rolling benchmark metrics
